use composition to remove code duplication with Account and Approval classes use class ContractValidationUtil instead of inheritance refactoring bug fixes

+++ src/main/java/de/fs92/defi/oasis/OasisOffer.java
@@ -1,20 +1,19 @@
 package de.fs92.defi.oasis;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Map;
 
 public class OasisOffer {
   public final BigInteger offerId;
-  public final Map<String, BigDecimal> offerValues;
-  public final BigDecimal bestOfferDaiPerEth;
-  public final BigDecimal profit;
+  public final Map<String, BigInteger> offerValues;
+  public final BigInteger bestOfferDaiPerEth;
+  public final BigInteger profit;
 
   public OasisOffer(
       BigInteger offerId,
-      Map<String, BigDecimal> offerValues,
-      BigDecimal bestOfferDaiPerEth,
-      BigDecimal profit) {
+      Map<String, BigInteger> offerValues,
+      BigInteger bestOfferDaiPerEth,
+      BigInteger profit) {
     this.offerId = offerId;
     this.offerValues = offerValues;
     this.bestOfferDaiPerEth = bestOfferDaiPerEth;
