+++ src/main/java/de/fs92/defi/uniswap/UniswapOffer.java
@@ -1,13 +1,12 @@
 package de.fs92.defi.uniswap;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public class UniswapOffer {
   public final BigInteger buyableAmount;
-  public final BigDecimal profit;
+  public final BigInteger profit;
 
-  public UniswapOffer(BigInteger buyableAmount, BigDecimal profit) {
+  public UniswapOffer(BigInteger buyableAmount, BigInteger profit) {
     this.buyableAmount = buyableAmount;
     this.profit = profit;
   }
